EP_GAN/requirements.txt
Flask
torch
torchvision
Pillow
numpy
scikit-image
